Create a Moodle plugin to connect to OpenVeo
This plugin is a local plugin which exposes the
OpenVeo REST PHP client to connect to OpenVeo API.
It also offers an administration page to configure
the connection to the OpenVeo web service. OpenVeo
REST PHP client library is not embedded in this
plugin and must be installed manually (see the
README.md file).

Other Moodle plugins requiring the use of OpenVeo
web servicen have to depend on this plugin.